+++ frontend/src/App.tsx
@@ -5,6 +5,7 @@ import BodyMetricsPage from './pages/BodyMetrics'
 import LabResultsPage from './pages/LabResults'
 import VitalSignsPage from './pages/VitalSigns'
 import SettingsPage from './pages/Settings'
+import ReferenceRangesPage from './pages/ReferenceRanges'
 
 export default function App() {
   return (
@@ -16,6 +17,7 @@ export default function App() {
           <Route path="body-metrics" element={<BodyMetricsPage />} />
           <Route path="lab-results" element={<LabResultsPage />} />
           <Route path="vital-signs" element={<VitalSignsPage />} />
+          <Route path="ranges" element={<ReferenceRangesPage />} />
           <Route path="settings" element={<SettingsPage />} />
         </Route>
       </Routes>